+++ tests/test.rs
@@ -364,21 +364,21 @@ fn dyn_type() {
     }
 }
 
-// #[test]
-// fn no_infer_outlives() {
-//     trait Bar<X> {
-//         type Y;
-//     }
+#[test]
+fn no_infer_outlives() {
+    trait Bar<X> {
+        type Y;
+    }
 
-//     struct Example<A>(A);
+    struct Example<A>(A);
 
-//     impl<X, T> Bar<X> for Example<T> {
-//         type Y = Option<T>;
-//     }
+    impl<X, T> Bar<X> for Example<T> {
+        type Y = Option<T>;
+    }
 
-//     pin_project! {
-//         struct Foo<A, B> {
-//             _x: <Example<A> as Bar<B>>::Y,
-//         }
-//     }
-// }
+    pin_project! {
+        struct Foo<A, B> {
+            _x: <Example<A> as Bar<B>>::Y,
+        }
+    }
+}