+++ src/lib.rs
@@ -70,7 +70,7 @@
 
 #![no_std]
 #![recursion_limit = "256"]
-#![doc(html_root_url = "https://docs.rs/pin-project-lite/0.1.6")]
+#![doc(html_root_url = "https://docs.rs/pin-project-lite/0.1.7")]
 #![doc(test(
     no_crate_inject,
     attr(deny(warnings, rust_2018_idioms, single_use_lifetimes), allow(dead_code))