+++ CHANGELOG.md
@@ -8,6 +8,8 @@ This project adheres to [Semantic Versioning](https://semver.org).
 
 * Fix lifetime inference error when an associated type is used in fields.
 
+* Suppress `clippy::used_underscore_binding` lint in generated code.
+
 ## [0.1.6] - 2020-05-31
 
 * [Support lifetime bounds in where clauses.][18]