+++ CHANGELOG.md
@@ -6,11 +6,13 @@ This project adheres to [Semantic Versioning](https://semver.org).
 
 ## [Unreleased]
 
-* Support ?Sized bounds in where clauses.
+* Support `?Sized` bounds in where clauses.
 
 * Fix lifetime inference error when an associated type is used in fields.
 
 * Suppress `clippy::used_underscore_binding` lint in generated code.
+
+* Documentation improvements.
 
 ## [0.1.6] - 2020-05-31